test: move execute_gmp_link_token to harness
